requirements.txt
# Core Flask packages
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==2.2.3
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0

# Server
gunicorn==21.2.0

# Document handling
python-docx==0.8.11
docxtpl==0.16.4

# Additional utilities
Jinja2==3.0.1
MarkupSafe>=2.1.1
itsdangerous==2.0.1
click==8.0.1

# Date handling
pytz==2021.1

# Form validation
WTForms==2.3.3
email-validator==1.1.3

# New packages
reportlab>=4.0.0

# Added for font handling
pillow==9.5.0
